handle http request without content-type
A well formed http request should have a `Content-type` header, but when
the content-type is not defined, it will cause our security middleware
to throw a keyerror.

This change makes sure that an error is not thrown in this case. When
there is no content-type header, the token will not be refreshed