generate tokens with HMAC instead of random strings, #25
- This makes it almost impossible for an attacker to forge a token,
even if they find a way to set a cookie.
It also provides a verified timestamp so we can easily check the token age.
In future, it might even be possible to eliminate the cookie altogether,
so long as sufficient information is contained in the token,
eg username + timestamp + intended action.