Refactor brute force lockout to use username key
This used to be specific to the IP address, but it was found that
brute force attacks could still be performed on an account by spoofing
the IP headers when logging in. Now the stored information about failed
login attempts is stored on a per username basis.

The username is guaranteed to be unique by the CKAN db definition... even
if a user has been disabled you can't create a new user with the same username.