Provide a way to detect which form is submitted
The serverside had no easy way to know when a user
is submitting just the username and password, or
they have already done that and they are on the 2fa
form. Now if the lockout occurs, we can lock the
user in the form they are supplying, not giving
any feedback that the account has been locked.